Get rid of package objects
